fix: minified react #425 에러 테스트
