+++ components/Card/Card.tsx
@@ -6,6 +6,8 @@ import classNames from "classnames/bind";
 import Image from "next/image";
 
 import { Folder, Link } from "@/utils/api/types";
+import calculatePassedTime from "@/utils/calculatePassedTime";
+import changeDateFormat from "@/utils/changeDateFormat";
 
 import styles from "./Card.module.scss";
 import SelectMenu from "./SelectMenu";
@@ -14,46 +16,6 @@ import Star from "./Star";
 import kebab from "@/public/images/show-more.png";
 
 const cx = classNames.bind(styles);
-
-const calculatePassedTime = (time: string): string => {
-  const MINUTE = 60 * 1000;
-  const HOUR = 60 * MINUTE;
-  const DAY = 24 * HOUR;
-  const MONTH = 31 * DAY;
-  const YEAR = 12 * MONTH;
-
-  const currentTime = new Date();
-  const createdTime = new Date(time);
-  const timeDiff = currentTime.getTime() - createdTime.getTime();
-
-  switch (true) {
-    case timeDiff < 2 * MINUTE:
-      return "1 minute ago";
-    case timeDiff < HOUR:
-      return `${Math.floor(timeDiff / MINUTE)} minutes ago`;
-    case timeDiff < 2 * HOUR:
-      return "1 hour ago";
-    case timeDiff < DAY:
-      return `${Math.floor(timeDiff / HOUR)} hours ago`;
-    case timeDiff < 2 * DAY:
-      return "1 day ago";
-    case timeDiff < MONTH:
-      return `${Math.floor(timeDiff / DAY)} days ago`;
-    case timeDiff < 2 * MONTH:
-      return "1 month ago";
-    case timeDiff < YEAR:
-      return `${Math.floor(timeDiff / MONTH)} months ago`;
-    case timeDiff < 2 * YEAR:
-      return "1 year ago";
-    default:
-      return `${Math.floor(timeDiff / YEAR)} years ago`;
-  }
-};
-
-const changeDateFormat = (time: string): string => {
-  return Intl.DateTimeFormat("kr").format(new Date(time));
-};
-
 interface CardProps {
   link: Link;
   folders?: Folder[];
