+++ utils/changeDateFormat.ts
@@ -1,0 +1,5 @@
+const changeDateFormat = (time: string): string => {
+  return Intl.DateTimeFormat("kr").format(new Date(time));
+};
+
+export default changeDateFormat;
